requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
pydantic-settings==2.10.1
email-validator==2.2.0
psycopg[binary]==3.3.3
pytest==8.4.2
httpx==0.28.1
